refactor: 컴포넌트 분리 및 메인페이지 리팩토링
