+++ app/lib/constans.ts
@@ -110,8 +110,8 @@ export const FilterableFields = [
   'Name',
   'CA',
   'PA',
-  'AP',
   'Salary',
+  'AP',
   'Age',
   'Wor',
   'One',
