+++ components/observer.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { IProps } from '@/app/type';
 import Image from 'next/image';
 import React, { useEffect, useRef } from 'react';
