+++ components/provider.tsx
@@ -1,0 +1,21 @@
+'use client';
+import { Provider } from 'react-redux';
+import { ThemeProvider } from '@material-tailwind/react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import store from '@/app/lib/store/store';
+
+const queryClient = new QueryClient();
+const CustomProvider = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <Provider store={store}>
+      <ThemeProvider>
+        <QueryClientProvider client={queryClient}>
+          {children}
+          <ReactQueryDevtools />
+        </QueryClientProvider>
+      </ThemeProvider>
+    </Provider>
+  );
+};
+export default CustomProvider;
