+++ tailwind.config.js
@@ -92,6 +92,12 @@ module.exports = withMT({
         '.horizontal-text': {
           writingMode: 'horizontal-tb',
         },
+        '.multi-line-truncate': {
+          display: '-webkit-box',
+          '-webkit-line-clamp': '2',
+          '-webkit-box-orient': 'vertical',
+          overflow: 'hidden',
+        },
       };
       addUtilities(newUtilities, ['responsive', 'hover']);
     },
